python: introduce APIRouter
