+++ webapp/python/src/isuride/app/sql.py
@@ -1,0 +1,3 @@
+import sqlalchemy
+
+engine = sqlalchemy.create_engine("mysql+pymysql://isucon:isucon@localhost/isuride")
